requirements.txt
numpy
scikit-learn
python-dotenv
pandas